Method that receives image has been moved to service. Changed Post to Get.
