+++ sandbox-web/src/main/java/com/exadel/sandbox/team5/controller/DiscountRestController.java
@@ -1,10 +1,10 @@
 package com.exadel.sandbox.team5.controller;
 
-import com.exadel.sandbox.team5.barcodes.QRCodeGenerator;
 import com.exadel.sandbox.team5.dto.DiscountDto;
 import com.exadel.sandbox.team5.entity.Review;
 import com.exadel.sandbox.team5.service.DiscountService;
 import com.exadel.sandbox.team5.service.ReviewService;
+import com.exadel.sandbox.team5.service.impl.DiscountServiceImpl;
 import io.swagger.annotations.ApiOperation;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
@@ -13,8 +13,6 @@ import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import javax.imageio.ImageIO;
-import java.io.ByteArrayOutputStream;
 import java.util.List;
 
 @RestController
@@ -25,6 +23,7 @@ public class DiscountRestController {
 
     private final DiscountService service;
     private final ReviewService reviewService;
+    private final DiscountServiceImpl discountService;
 
     @GetMapping("/{id}")
     public DiscountDto getDiscount(@PathVariable Long id) {
@@ -58,17 +57,10 @@ public class DiscountRestController {
     }
 
     @ApiOperation("Generating QR code with text \"Exadel employee. Special discount\"")
-    @PostMapping(value = "/qrcode", produces = MediaType.IMAGE_PNG_VALUE)
+    @GetMapping(value = "/qrcode", produces = MediaType.IMAGE_PNG_VALUE)
     @ResponseStatus(HttpStatus.OK)
     public ResponseEntity<byte[]> generateQRCode() {
-        try (var baos = new ByteArrayOutputStream()) {
-            var image = QRCodeGenerator.generateQRCodeImage("Exadel employee. Special discount");
-            ImageIO.write(image, "png", baos);
-            return ResponseEntity.ok(baos.toByteArray());
-        } catch (Exception e) {
-            log.error("There was an error during barcode generation", e);
-            return ResponseEntity.badRequest().build();
-        }
+        return ResponseEntity.ok(discountService.generateQRCode());
     }
 }
